requirements.txt
streamlit>=1.28.0
openai>=1.3.0
newsapi-python>=0.2.7
gnews>=0.2.9
spacy>=3.7.0
dateparser>=1.2.0
plotly>=5.17.0
pandas>=2.1.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
google-generativeai>=0.3.0

